tools/requirements.txt
easyocr
numpy<2
opencv-python
pdfplumber
psycopg2-binary
openai
requests
